# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "Analysis": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus",
            "default": "draft"
          },
          "title": {
            "maxLength": 200,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          },
          "trial": {
            "$ref": "#/components/schemas/TrialDraft"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": ["owner_id", "title", "trial"],
        "title": "Analysis",
        "type": "object"
      },
      "AnalysisStatus": {
        "enum": ["draft", "ready"],
        "title": "AnalysisStatus",
        "type": "string"
      },
      "CreateAnalysisRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "trial": {
            "$ref": "#/components/schemas/TrialDraft"
          }
        },
        "required": ["trial"],
        "title": "CreateAnalysisRequest",
        "type": "object"
      },
      "Endpoint": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "measure": {
            "title": "Measure",
            "type": "string"
          },
          "time_frame": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Frame"
          }
        },
        "required": ["measure"],
        "title": "Endpoint",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Intervention": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "intervention_type": {
            "title": "Intervention Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["name", "intervention_type"],
        "title": "Intervention",
        "type": "object"
      },
      "Site": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "facility": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": ["country"],
        "title": "Site",
        "type": "object"
      },
      "SourceReference": {
        "properties": {
          "field_locators": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Field Locators",
            "type": "object"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "record_id": {
            "title": "Record Id",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Version"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": ["provider", "record_id", "url", "retrieved_at"],
        "title": "SourceReference",
        "type": "object"
      },
      "StudyDesign": {
        "properties": {
          "allocation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allocation"
          },
          "arm_count": {
            "default": 0,
            "title": "Arm Count",
            "type": "integer"
          },
          "intervention_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervention Model"
          },
          "masking": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Masking"
          },
          "primary_purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Purpose"
          },
          "study_type": {
            "title": "Study Type",
            "type": "string"
          }
        },
        "required": ["study_type"],
        "title": "StudyDesign",
        "type": "object"
      },
      "TrialDraft": {
        "properties": {
          "biomarker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Biomarker"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Conditions",
            "type": "array"
          },
          "eligibility_criteria": {
            "minLength": 3,
            "title": "Eligibility Criteria",
            "type": "string"
          },
          "enrollment": {
            "minimum": 1.0,
            "title": "Enrollment",
            "type": "integer"
          },
          "enrollment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrollment Type"
          },
          "healthy_volunteers": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy Volunteers"
          },
          "indication": {
            "minLength": 2,
            "title": "Indication",
            "type": "string"
          },
          "interventions": {
            "items": {
              "$ref": "#/components/schemas/Intervention"
            },
            "minItems": 1,
            "title": "Interventions",
            "type": "array"
          },
          "maximum_age": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum Age"
          },
          "minimum_age": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Age"
          },
          "nct_id": {
            "pattern": "^NCT\\d{8}$",
            "title": "Nct Id",
            "type": "string"
          },
          "overall_status": {
            "title": "Overall Status",
            "type": "string"
          },
          "phases": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Phases",
            "type": "array"
          },
          "primary_completion_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Completion Date"
          },
          "primary_endpoints": {
            "items": {
              "$ref": "#/components/schemas/Endpoint"
            },
            "minItems": 1,
            "title": "Primary Endpoints",
            "type": "array"
          },
          "secondary_endpoints": {
            "items": {
              "$ref": "#/components/schemas/Endpoint"
            },
            "title": "Secondary Endpoints",
            "type": "array"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "title": "Sites",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/SourceReference"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "study_design": {
            "$ref": "#/components/schemas/StudyDesign"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "maxLength": 600,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "nct_id",
          "title",
          "overall_status",
          "indication",
          "conditions",
          "phases",
          "interventions",
          "eligibility_criteria",
          "study_design",
          "primary_endpoints",
          "enrollment",
          "source"
        ],
        "title": "TrialDraft",
        "type": "object"
      },
      "TrialSearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrialSearchResult"
            },
            "title": "Items",
            "type": "array"
          },
          "next_page_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page Token"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "required": ["items"],
        "title": "TrialSearchResponse",
        "type": "object"
      },
      "TrialSearchResult": {
        "properties": {
          "conditions": {
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "enrollment": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrollment"
          },
          "interventions": {
            "items": {
              "type": "string"
            },
            "title": "Interventions",
            "type": "array"
          },
          "nct_id": {
            "title": "Nct Id",
            "type": "string"
          },
          "overall_status": {
            "title": "Overall Status",
            "type": "string"
          },
          "phases": {
            "items": {
              "type": "string"
            },
            "title": "Phases",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "us_site_count": {
            "default": 0,
            "title": "Us Site Count",
            "type": "integer"
          }
        },
        "required": [
          "nct_id",
          "title",
          "overall_status",
          "phases",
          "conditions",
          "interventions"
        ],
        "title": "TrialSearchResult",
        "type": "object"
      },
      "UpdateAnalysisRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "trial": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrialDraft"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UpdateAnalysisRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Illustrative decision-support API. Not a validated clinical or enrollment-prediction system.",
    "title": "Oncology Trial Feasibility Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Live Health Live Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live",
        "tags": ["health"]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready Health Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": ["health"]
      }
    },
    "/v1/analyses": {
      "get": {
        "operationId": "list_analyses_v1_analyses_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Analysis"
                  },
                  "title": "Response List Analyses V1 Analyses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Analyses",
        "tags": ["analyses"]
      },
      "post": {
        "operationId": "create_analysis_v1_analyses_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Analysis",
        "tags": ["analyses"]
      }
    },
    "/v1/analyses/{analysis_id}": {
      "delete": {
        "operationId": "delete_analysis_v1_analyses__analysis_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Analysis",
        "tags": ["analyses"]
      },
      "get": {
        "operationId": "get_analysis_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis",
        "tags": ["analyses"]
      },
      "patch": {
        "operationId": "update_analysis_v1_analyses__analysis_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Analysis",
        "tags": ["analyses"]
      }
    },
    "/v1/trials": {
      "get": {
        "operationId": "search_trials_v1_trials_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 2,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Trials",
        "tags": ["trials"]
      }
    },
    "/v1/trials/{nct_id}": {
      "get": {
        "operationId": "get_trial_v1_trials__nct_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "nct_id",
            "required": true,
            "schema": {
              "title": "Nct Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trial",
        "tags": ["trials"]
      }
    }
  }
}